super1: remove all dlm lock stuffs
Besides the three funcs, there are some more funcs
which access superblock should be protected by dlm
lock.

Instead of call the lock funcs inside each funcs,
it is better to protect it from higher level, so
remove them first, and provide the protection in
mdadm.c which will be changed in next patch.

Signed-off-by: Guoqing Jiang <gqjiang@suse.com>