util: improve the dlm protection for clustered raid
Some places for dlm need to be improved such as:

1. replace ls_unlock with ls_unlock_wait since we
should return when unlock operation is complete.

2. inspired by lvm, let's also try to use the existed
lockspace first before creat a lockspace blindly if
the lockspace not released for some reason.

3. try more times before quit if EAGAIN happened for
locking.

Then lock_cluster is introduced which will be called
from main(), it tries to hold mutex before get dlm
lock to make the test script happy, because the test
script launches background tasks. And unlock_cluster
does the opposite thing.

Signed-off-by: Guoqing Jiang <gqjiang@suse.com>