mdadm: protect clustered raid till mdadm exits
Previously, the dlm locking only protects several
functions which writes to superblock (update_super,
add_to_super and store_super), and we missed other
funcs such as add_internal_bitmap.

We also need to call the funcs which read superblock
under the locking protection to avoid consistent
issue.

So instead of locking from each funcs which access
superblock, it is better move the locking mechanism
to the main() function.

We can identify it is a clustered raid or not according
to different conditions based on the different mode.
So the change should not have effect on non-clustered
array.

Signed-off-by: Guoqing Jiang <gqjiang@suse.com>